requirements.txt
streamlit==1.26.0
ultralytics>=8.1.3
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python-headless>=4.7.0
Pillow<10
huggingface-hub>=0.20.0

